preview-api/pyproject.toml
[project]
name = "livrome-preview-api"
version = "0.1.0"
description = "Livrome preview generation API — /generate-page endpoint backed by Imagen 4 via Gemini Enterprise OAuth."
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "httpx>=0.27",
  "google-auth>=2.29",
  "google-cloud-secret-manager>=2.20",
  "python-multipart>=0.0.9",
  "pillow>=10.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "types-requests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]
